Added additional example of custom themed notification to Sample project
