+++ Notifications.Wpf.Core.Sample/CustomThemedNotificationContent.cs
@@ -1,0 +1,6 @@
+﻿namespace Notifications.Wpf.Core.Sample
+{
+    public class CustomThemedNotificationContent : NotificationContent
+    {
+    }
+}
